Address PR suggestions
* Move storage logic to separate package
* Encode now takes a prefix as first argument
* Decode now returns a prefix as well as its error
* Renamed prefixes and functions
* All indexes are compressed
