+++ service/state/commit.go
@@ -3,6 +3,7 @@ package state
 import (
 	"fmt"
 
+	"github.com/awfm9/flow-dps/service/state/storage"
 	"github.com/dgraph-io/badger/v2"
 
 	"github.com/onflow/flow-go/model/flow"
@@ -15,7 +16,7 @@ type Commit struct {
 func (c *Commit) ForHeight(height uint64) (flow.StateCommitment, error) {
 	var commit flow.StateCommitment
 	err := c.core.db.View(func(tx *badger.Txn) error {
-		return RetrieveCommit(height, &commit)(tx)
+		return storage.RetrieveCommitByHeight(height, &commit)(tx)
 	})
 	if err != nil {
 		return nil, fmt.Errorf("could not look up height: %w", err)
